refactor(ethereum): split tBTC chain adapter into per-concern files
pkg/chain/ethereum/tbtc.go had grown to ~2400 lines and ~90 methods
spanning every tBTC concern. Split it into per-concern files within the
same package so the adapter is navigable:

- tbtc.go: TbtcChain struct, constructor, shared constants
- tbtc_sortition.go: sortition pool, operator status, group selection
- tbtc_dkg.go: DKG lifecycle, result assembly and validation
- tbtc_inactivity.go: inactivity claims
- tbtc_deposit.go: deposit reveal, sweep proposals and proofs
- tbtc_redemption.go: redemption requests, proposals and proofs
- tbtc_wallet.go: wallet registration/state, heartbeat proposals
- tbtc_moving_funds.go: moving funds and moved funds sweep

Pure relocation: no declaration bodies changed. Verified by comparing
the AST of every top-level declaration before and after the split - all
96 declarations are byte-identical.
