+++ pkg/chain/ethereum/tbtc_sortition.go
@@ -1,0 +1,247 @@
+package ethereum
+
+import (
+	"context"
+	"fmt"
+	"math/big"
+
+	"github.com/ethereum/go-ethereum/common"
+
+	"github.com/keep-network/keep-core/pkg/chain"
+	"github.com/keep-network/keep-core/pkg/operator"
+	"github.com/keep-network/keep-core/pkg/tbtc"
+)
+
+// EcdsaWalletGroupParametersFromChain mirrors EcdsaDkgValidator sizing constants
+// when EcdsaDkgValidator contract address was configured under [ethereum]
+// contract addresses or developer.ecdsaDkgValidatorAddress alias. When absent,
+// returns (nil, nil) and callers use defaultGroupParameters(network).
+func (tc *TbtcChain) EcdsaWalletGroupParametersFromChain(
+	ctx context.Context,
+) (*tbtc.GroupParameters, error) {
+	if tc.ecdsaDkgValidatorAddress == (common.Address{}) {
+		return nil, nil
+	}
+	return ecdsaWalletGroupParametersFromValidator(
+		ctx,
+		tc.baseChain.client,
+		tc.ecdsaDkgValidatorAddress,
+	)
+}
+
+// Staking returns address of the TokenStaking contract the WalletRegistry is
+// connected to.
+func (tc *TbtcChain) Staking() (chain.Address, error) {
+	stakingContractAddress, err := tc.walletRegistry.Staking()
+	if err != nil {
+		return "", fmt.Errorf(
+			"failed to get the token staking address: [%w]",
+			err,
+		)
+	}
+
+	return chain.Address(stakingContractAddress.String()), nil
+}
+
+// IsRecognized checks whether the given operator is recognized by the TbtcChain
+// as eligible to join the network. If the operator has a stake delegation or
+// had a stake delegation in the past, it will be recognized.
+func (tc *TbtcChain) IsRecognized(operatorPublicKey *operator.PublicKey) (bool, error) {
+	operatorAddress, err := operatorPublicKeyToChainAddress(operatorPublicKey)
+	if err != nil {
+		return false, fmt.Errorf(
+			"cannot convert from operator key to chain address: [%v]",
+			err,
+		)
+	}
+
+	stakingProvider, err := tc.walletRegistry.OperatorToStakingProvider(
+		operatorAddress,
+	)
+	if err != nil {
+		return false, fmt.Errorf(
+			"failed to map operator [%v] to a staking provider: [%v]",
+			operatorAddress,
+			err,
+		)
+	}
+
+	if (stakingProvider == common.Address{}) {
+		return false, nil
+	}
+
+	// Check if the staking provider has an owner. This check ensures that there
+	// is/was a stake delegation for the given staking provider.
+	_, _, _, hasStakeDelegation, err := tc.baseChain.RolesOf(
+		chain.Address(stakingProvider.Hex()),
+	)
+	if err != nil {
+		return false, fmt.Errorf(
+			"failed to check stake delegation for staking provider [%v]: [%v]",
+			stakingProvider,
+			err,
+		)
+	}
+
+	if !hasStakeDelegation {
+		return false, nil
+	}
+
+	return true, nil
+}
+
+// OperatorToStakingProvider returns the staking provider address for the
+// operator. If the staking provider has not been registered for the
+// operator, the returned address is empty and the boolean flag is set to
+// false. If the staking provider has been registered, the address is not
+// empty and the boolean flag indicates true.
+func (tc *TbtcChain) OperatorToStakingProvider() (chain.Address, bool, error) {
+	stakingProvider, err := tc.walletRegistry.OperatorToStakingProvider(tc.key.Address)
+	if err != nil {
+		return "", false, fmt.Errorf(
+			"failed to map operator [%v] to a staking provider: [%v]",
+			tc.key.Address,
+			err,
+		)
+	}
+
+	if (stakingProvider == common.Address{}) {
+		return "", false, nil
+	}
+
+	return chain.Address(stakingProvider.Hex()), true, nil
+}
+
+// EligibleStake returns the current value of the staking provider's
+// eligible stake. Eligible stake is defined as the currently authorized
+// stake minus the pending authorization decrease. Eligible stake
+// is what is used for operator's weight in the sortition pool.
+// If the authorized stake minus the pending authorization decrease
+// is below the minimum authorization, eligible stake is 0.
+func (tc *TbtcChain) EligibleStake(stakingProvider chain.Address) (*big.Int, error) {
+	eligibleStake, err := tc.walletRegistry.EligibleStake(
+		common.HexToAddress(stakingProvider.String()),
+	)
+	if err != nil {
+		return nil, fmt.Errorf(
+			"failed to get eligible stake for staking provider %s: [%w]",
+			stakingProvider,
+			err,
+		)
+	}
+
+	return eligibleStake, nil
+}
+
+// IsPoolLocked returns true if the sortition pool is locked and no state
+// changes are allowed.
+func (tc *TbtcChain) IsPoolLocked() (bool, error) {
+	return tc.sortitionPool.IsLocked()
+}
+
+// IsOperatorInPool returns true if the operator is registered in
+// the sortition pool.
+func (tc *TbtcChain) IsOperatorInPool() (bool, error) {
+	return tc.walletRegistry.IsOperatorInPool(tc.key.Address)
+}
+
+// IsOperatorUpToDate checks if the operator's authorized stake is in sync
+// with operator's weight in the sortition pool.
+// If the operator's authorized stake is not in sync with sortition pool
+// weight, function returns false.
+// If the operator is not in the sortition pool and their authorized stake
+// is non-zero, function returns false.
+func (tc *TbtcChain) IsOperatorUpToDate() (bool, error) {
+	return tc.walletRegistry.IsOperatorUpToDate(tc.key.Address)
+}
+
+// JoinSortitionPool executes a transaction to have the operator join the
+// sortition pool.
+func (tc *TbtcChain) JoinSortitionPool() error {
+	_, err := tc.walletRegistry.JoinSortitionPool()
+	return err
+}
+
+// UpdateOperatorStatus executes a transaction to update the operator's
+// state in the sortition pool.
+func (tc *TbtcChain) UpdateOperatorStatus() error {
+	_, err := tc.walletRegistry.UpdateOperatorStatus(tc.key.Address)
+	return err
+}
+
+// IsEligibleForRewards checks whether the operator is eligible for rewards
+// or not.
+func (tc *TbtcChain) IsEligibleForRewards() (bool, error) {
+	return tc.sortitionPool.IsEligibleForRewards(tc.key.Address)
+}
+
+// Checks whether the operator is able to restore their eligibility for
+// rewards right away.
+func (tc *TbtcChain) CanRestoreRewardEligibility() (bool, error) {
+	return tc.sortitionPool.CanRestoreRewardEligibility(tc.key.Address)
+}
+
+// Restores reward eligibility for the operator.
+func (tc *TbtcChain) RestoreRewardEligibility() error {
+	_, err := tc.sortitionPool.RestoreRewardEligibility(tc.key.Address)
+	return err
+}
+
+// Returns true if the chaosnet phase is active, false otherwise.
+func (tc *TbtcChain) IsChaosnetActive() (bool, error) {
+	return tc.sortitionPool.IsChaosnetActive()
+}
+
+// Returns true if operator is a beta operator, false otherwise.
+// Chaosnet status does not matter.
+func (tc *TbtcChain) IsBetaOperator() (bool, error) {
+	return tc.sortitionPool.IsBetaOperator(tc.key.Address)
+}
+
+// GetOperatorID returns the ID number of the given operator address. An ID
+// number of 0 means the operator has not been allocated an ID number yet.
+func (tc *TbtcChain) GetOperatorID(
+	operatorAddress chain.Address,
+) (chain.OperatorID, error) {
+	return tc.sortitionPool.GetOperatorID(
+		common.HexToAddress(operatorAddress.String()),
+	)
+}
+
+// SelectGroup returns the group members selected for the current group
+// selection. The function returns an error if the chain's state does not allow
+// for group selection at the moment.
+func (tc *TbtcChain) SelectGroup() (*tbtc.GroupSelectionResult, error) {
+	operatorsIDs, err := tc.walletRegistry.SelectGroup()
+	if err != nil {
+		return nil, fmt.Errorf(
+			"cannot select group in the sortition pool: [%v]",
+			err,
+		)
+	}
+
+	operatorsAddresses, err := tc.sortitionPool.GetIDOperators(operatorsIDs)
+	if err != nil {
+		return nil, fmt.Errorf(
+			"cannot convert operators' IDs to addresses: [%v]",
+			err,
+		)
+	}
+
+	// Should not happen as this is guaranteed by the contract but, just in case.
+	if len(operatorsIDs) != len(operatorsAddresses) {
+		return nil, fmt.Errorf("operators IDs and addresses mismatch")
+	}
+
+	ids := make([]chain.OperatorID, len(operatorsIDs))
+	addresses := make([]chain.Address, len(operatorsIDs))
+	for i := range ids {
+		ids[i] = operatorsIDs[i]
+		addresses[i] = chain.Address(operatorsAddresses[i].String())
+	}
+
+	return &tbtc.GroupSelectionResult{
+		OperatorsIDs:       ids,
+		OperatorsAddresses: addresses,
+	}, nil
+}
